Created kappa_d eval. for uhecr dataset
