+++ fancy/interfaces/__init__.py
@@ -1,2 +1,3 @@
 from .data import *
+from .uhecr import *
 from .integration import *
